# app-react/openapi.json
{"openapi":"3.0.1","info":{"title":"API Cave à vin","description":"Documentation OpenAPI des endpoints /api","contact":{"name":"Support","email":"support@example.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/wines/{id}":{"put":{"tags":["wine-controller"],"operationId":"updateWine","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WineDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Wine"}}}}}},"delete":{"tags":["wine-controller"],"operationId":"deleteWine","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/users/current":{"get":{"tags":["user-controller"],"operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateCurrentUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/caves/{id}":{"get":{"tags":["cave-controller"],"operationId":"getCaveById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CaveDto"}}}}}},"put":{"tags":["cave-controller"],"operationId":"updateCave","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaveDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CaveDto"}}}}}},"delete":{"tags":["cave-controller"],"operationId":"deleteCave","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/wines":{"post":{"tags":["wine-controller"],"operationId":"createWine","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WineDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WineDto"}}}}}}},"/api/users/register":{"post":{"tags":["user-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/login":{"post":{"tags":["user-controller"],"operationId":"createAuthenticationToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/lignes":{"post":{"tags":["ligne-controller"],"operationId":"createLigne","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LigneDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LigneDto"}}}}}}},"/api/caves":{"get":{"tags":["cave-controller"],"operationId":"getAllCaves","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CaveDto"}}}}}}},"post":{"tags":["cave-controller"],"operationId":"createCave","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaveDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CaveDto"}}}}}}},"/api/cave-management/{caveId}/add-casier":{"post":{"tags":["cave-management-controller"],"operationId":"ajouterCasier","parameters":[{"name":"caveId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"numeroCasier","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CasierDto"}}}}}}},"/api/cave-management/swap-wines":{"post":{"tags":["cave-management-controller"],"operationId":"swapWines","parameters":[{"name":"wineId1","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"wineId2","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/casiers":{"post":{"tags":["casier-controller"],"operationId":"createCasier","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CasierDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CasierDto"}}}}}}},"/api/casiers/cave/{caveId}":{"get":{"tags":["casier-controller"],"operationId":"getCasiersByCaveId","parameters":[{"name":"caveId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CasierDto"}}}}}}}},"/api/casiers/{id}":{"delete":{"tags":["casier-controller"],"operationId":"deleteCasier","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"WineDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"chateau":{"type":"string"},"appellation":{"type":"string"},"taille":{"type":"string"},"annee":{"type":"integer","format":"int32"},"couleur":{"type":"string","enum":["ROUGE","BLANC","ROSE","PETILLANT","CHAMPAGNE"]},"prix":{"type":"number"},"ligneId":{"type":"integer","format":"int64"},"position":{"type":"integer","format":"int32"}}},"Casier":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"numeroCasier":{"type":"integer","format":"int32"},"position":{"type":"integer","format":"int32"},"lignes":{"type":"array","items":{"$ref":"#/components/schemas/Ligne"}},"cave":{"$ref":"#/components/schemas/Cave"},"nombreTotalBouteilles":{"type":"integer","format":"int32"},"capaciteMaximale":{"type":"integer","format":"int32"}}},"Cave":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"description":{"type":"string"},"nombreLignesParCasier":{"type":"integer","format":"int32"},"capacitesParLigne":{"type":"array","items":{"type":"integer","format":"int32"}},"casiers":{"type":"array","items":{"$ref":"#/components/schemas/Casier"}},"owner":{"$ref":"#/components/schemas/User"},"nombreTotalBouteilles":{"type":"integer","format":"int32"},"capaciteTotaleCave":{"type":"integer","format":"int32"}}},"Ligne":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombreBouteillesMax":{"type":"integer","format":"int32"},"position":{"type":"integer","format":"int32"},"bouteilles":{"type":"array","items":{"$ref":"#/components/schemas/Wine"}},"casier":{"$ref":"#/components/schemas/Casier"},"nombreBouteillesActuelles":{"type":"integer","format":"int32"},"pleine":{"type":"boolean"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"caves":{"type":"array","items":{"$ref":"#/components/schemas/Cave"}},"fullName":{"type":"string"}}},"Wine":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"chateau":{"type":"string"},"appellation":{"type":"string"},"annee":{"type":"integer","format":"int32"},"position":{"type":"integer","format":"int32"},"taille":{"type":"string"},"couleur":{"type":"string","enum":["ROUGE","BLANC","ROSE","PETILLANT","CHAMPAGNE"]},"prix":{"type":"number"},"ligne":{"$ref":"#/components/schemas/Ligne"}}},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}}},"CasierDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"numeroCasier":{"type":"integer","format":"int32"},"caveId":{"type":"integer","format":"int64"},"position":{"type":"integer","format":"int32"},"lignes":{"type":"array","items":{"$ref":"#/components/schemas/LigneDto"}}}},"CaveDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"description":{"type":"string"},"nombreLignesParCasier":{"type":"integer","format":"int32"},"capacitesParLigne":{"type":"array","items":{"type":"integer","format":"int32"}},"ownerId":{"type":"integer","format":"int64"},"casiers":{"type":"array","items":{"$ref":"#/components/schemas/CasierDto"}}}},"LigneDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombreBouteillesMax":{"type":"integer","format":"int32"},"casierId":{"type":"integer","format":"int64"},"position":{"type":"integer","format":"int32"},"bouteilles":{"type":"array","items":{"$ref":"#/components/schemas/WineDto"}}}},"UserRegistrationDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}}}}}}